[backported/clang-9][preamble] Stop circular inclusion of main file when building preamble
https://reviews.llvm.org/D53866

If a header file was processed for the second time, we could end up with
a wrong conditional stack and skipped ranges:

In the particular example, if the header guard is evaluated the second
time and it is decided to skip the conditional block, the corresponding
"#endif" is never seen since the preamble does not include it and we end
up in the Tok.is(tok::eof) case with a wrong conditional stack.

Detect the circular inclusion, emit a diagnostic and stop processing the
inclusion.

Cherry picked from commit
clang.git/c12b012bb7465299490cf93c2ae90499a5c417d5.

Fixes: QTCREATORBUG-20883
Change-Id: I75152562273b34cfc474ed380006f5c7f42ed6e1
Reviewed-by: Marco Bubke <marco.bubke@qt.io>