Fix performance issue on macOS
From the getpriority manpage: "When a thread or process is in a
background state the scheduling priority is set to the lowest value,
disk IO is throttled, and network IO is throttled for any sockets opened
after going into background state.".

The IO throttling when putting threads into background state absolutely
kills the indexing performance, and makes the multi-threading
ineffective. Independent of the number of working threads, the actual
performance was reduced to something similar to a single thread.

Instead of putting the indexing threads into background state, just set
their scheduling priority to the minimum.

Change-Id: Id5069df9c88ff1c00d2f6810fcddd4cff745feeb
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>